requirements.txt
flask
openai
Flask-Session
azure-data-tables
xlsxwriter